move provider and registry implementations to volume.impl package
